Picklist split to seperate comp
